feat: đề cương (#37)
* feat: load hoc phan

* feat: them,sua,xoa de cuong

+++ src/main/java/com/QLDaoTao/dto/response/DeCuongChiTietResponse.java
@@ -4,7 +4,7 @@ import com.QLDaoTao.model.DeCuongChiTiet;
 
 public record DeCuongChiTietResponse(
         Long deCuongId,
-        Long maHp,
+        Integer maHp,
         String tenBoPhan,
         String diemDanhGia,
         Float trongSo,
